requirements.txt
yt-dlp=2025.3.31
aiogram=3.19.0
python_dotenv=1.1.0
asyncio=3.4.3
